+++ build.sc
@@ -1,11 +1,7 @@
-import mill._
-import mill.scalalib._
-import mill.scalalib.publish._
-import coursier.maven.MavenRepository
 import $file.common
 import $file.hardfloat.build
 import $file.`api-config-chipsalliance`.`build-rules`.mill.build
-import mill.define.Sources
+import os.{Path, proc, resource}
 
 object configRocket extends `api-config-chipsalliance`.`build-rules`.mill.build.config with PublishModule {
   override def millSourcePath = os.pwd / "api-config-chipsalliance" / "design" / "craft"
@@ -36,47 +32,20 @@ object hardfloatRocket extends hardfloat.build.hardfloat {
   ) else Agg.empty[Dep]
 }
 
-object sanitytests extends ScalaModule with TestModule.Utest {
-  override def scalaVersion = T {
-    rocketchip.scalaVersion()
-  }
-
-  override def moduleDeps: Seq[JavaModule] = Seq(rocketchip)
-
-  override def scalacPluginClasspath = T { super.scalacPluginClasspath() }
-
-  object ivys {
-    val chisel3 = ivy"edu.berkeley.cs::chisel3:3.5.3"
-    val firrtl = ivy"edu.berkeley.cs::firrtl_2.13:1.5.3"
-    val scalatest = ivy"org.scalatest::scalatest:3.2.0"
-    val oslib = ivy"com.lihaoyi::os-lib:0.7.8"
-    val utest = ivy"com.lihaoyi::utest:0.7.10"
-  }
-
-  override def ivyDeps = Agg(
-    ivys.chisel3,
-    ivys.scalatest,
-    ivys.oslib,
-    ivys.utest
-  )
-
-  object helper {
-    val isMac = System.getProperty("os.name").toLowerCase.startsWith("mac")
-  }
-
-  def libraryResources = T.sources {
-    //    if (!helper.isMac) {
-    //      val x86Dir = T.ctx.dest
-    //      val riscv64Dir = T.ctx.dest / "riscv64"
-    //      os.proc("make", s"DESTDIR=${x86Dir}", "install").call(spike.compile())
-    //      os.proc("make", s"DESTDIR=${riscv64Dir}", "install").call(compilerrt.compile())
-    //      os.proc("make", s"DESTDIR=${riscv64Dir}", "install").call(musl.compile())
-    //      os.copy.into(pk.compile(), riscv64Dir)
-    //    }
+object spike extends Module {
+  override def millSourcePath = os.pwd / "riscv-isa-sim"
+  def compile = T.persistent {
+    os.proc(millSourcePath / "configure", "--prefix", "/usr", "--without-boost", "--without-boost-asio", "--without-boost-regex").call(
+      T.ctx.dest, Map(
+        "CC" -> "clang",
+        "CXX" -> "clang++",
+        "AR" -> "llvm-ar",
+        "RANLIB" -> "llvm-ranlib",
+        "LD" -> "lld",
+      )
+    )
+    os.proc("make", "-j", Runtime.getRuntime().availableProcessors()).call(T.ctx.dest)
     T.ctx.dest
-  }
-  override def resources: Sources = T.sources {
-    super.resources() ++ libraryResources()
   }
 }
 
@@ -93,5 +62,76 @@ object rocketchip extends common.CommonRocketChip {
 
   def configModule = configRocket
 
-  def testModule = sanitytests
+  object sanitytests extends ScalaModule {
+    override def scalaVersion = T {
+      rocketchip.scalaVersion()
+    }
+
+    override def moduleDeps: Seq[JavaModule] = Seq(rocketchip)
+
+    override def scalacPluginClasspath = T { super.scalacPluginClasspath() }
+
+    def resource(file: String): Path = Path(java.nio.file.Paths.get(getClass().getClassLoader().getResource(file).toURI))
+
+    def libraryResources = T.sources {
+      val x86Dir = T.ctx.dest
+      // build spike
+      os.proc("make", s"DESTDIR=${x86Dir}", "install").call(spike.compile())
+      // build hello world
+      val outputDirectory = os.pwd / "out" / "SanityTest"
+      os.remove.all(outputDirectory)
+      os.makeDir(outputDirectory)
+      os.proc(
+        "clang",
+        "-o", "hello",
+        s"${resource("csrc/hello.c")}",
+        "--target=riscv64",
+        "-mno-relax",
+        "-nostdinc",
+        s"-I${resource("riscv64/usr/include")}",
+        "-fuse-ld=lld",
+        "-nostdlib",
+        s"${resource("riscv64/usr/lib/crt1.o")}",
+        s"${resource("riscv64/usr/lib/crti.o")}",
+        s"${resource("riscv64/usr/lib/riscv64/libclang_rt.builtins-riscv64.a")}",
+        s"${resource("riscv64/usr/lib/libc.a")}",
+        s"${resource("riscv64/usr/lib/crtn.o")}",
+        "-static",
+      ).call(outputDirectory)
+      // build bootrom
+      val tmp = os.temp.dir()
+      val elf = tmp / "bootrom.elf"
+      val bin = tmp / "bootrom.bin"
+      val img = tmp / "bootrom.img"
+      // format: off
+      os.proc(
+        "clang",
+        "--target=riscv64", "-march=rv64gc",
+        "-mno-relax",
+        "-static",
+        "-nostdlib",
+        "-Wl,--no-gc-sections",
+        "-fuse-ld=lld", s"-T${resource("linker.ld")}",
+        s"${resource("bootrom.S")}",
+        "-o", elf
+      ).call()
+      os.proc(
+        "llvm-objcopy",
+        "-O", "binary",
+        elf,
+        bin
+      ).call()
+      os.proc(
+        "dd",
+        s"if=$bin",
+        s"of=$img",
+        "bs=128",
+        "count=1"
+      ).call()
+      T.ctx.dest
+    }
+    override def resources: Sources = T.sources {
+      super.resources() ++ libraryResources()
+    }
+  }
 }
