+++ src/test/scala/sanitytests/resources/csrc/comlog.cc
