+++ src/test/scala/sanitytests/resources/csrc/emulator.cc
