+++ src/test/scala/sanitytests/resources/csrc/remote_bitbang.cc
