+++ src/test/scala/sanitytests/src/VerilatorTest.scala
@@ -1,0 +1,28 @@
+package sanitytests
+
+import org.scalatest.funsuite.AnyFunSuite
+
+/** software dependencies:
+  * clang -> bootrom cross compiling / veriltor C compiling
+  * verilator -> emulator generation
+  * cmake -> simulation
+  * ninja -> fast verilator build tool
+  * spike -> isa behavior model linking in emulator
+  */
+object VerilatorTest extends AnyFunSuite {
+  val outputDirectory = os.pwd / "out" / "SanityTest"
+  os.remove.all(outputDirectory)
+  os.makeDir(outputDirectory)
+  test("build TestHarness emulator") {
+    val testHarness = classOf[freechips.rocketchip.system.TestHarness]
+    val configs = Seq(classOf[TestConfig], classOf[freechips.rocketchip.system.DefaultConfig])
+    val emulator = TestHarness(testHarness, configs, Some(outputDirectory)).emulator
+    test("Hello World!") {
+      os.proc(
+        s"$emulator",
+        s"${resource("riscv64/pk")}",
+        "hello",
+      ).call(outputDirectory)
+    }
+  }
+}
